Moved upgrade_to_supercarrier logic from tgogenerator.py to controlpoint.py
