+++ game/missiongenerator/tgogenerator.py
@@ -17,15 +17,7 @@ from dcs.action import DoScript, SceneryDestructionZone
 from dcs.condition import MapObjectIsDead
 from dcs.country import Country
 from dcs.point import StaticPoint
-from dcs.ships import (
-    CVN_71,
-    CVN_72,
-    CVN_73,
-    CVN_75,
-    CV_1143_5,
-    KUZNECOW,
-    Stennis,
-)
+
 from dcs.statics import Fortification
 from dcs.task import (
     ActivateBeaconCommand,
@@ -375,10 +367,7 @@ class GenericCarrierGenerator(GroundObjectGenerator):
                 )
 
     def get_carrier_type(self, group: TheaterGroup) -> Type[ShipType]:
-        carrier_type = group.units[0].type
-        if issubclass(carrier_type, ShipType):
-            return carrier_type
-        raise RuntimeError(f"First unit of TGO {group.name} is no Ship")
+        return self.control_point.get_carrier_group_type()
 
     def steam_into_wind(self, group: ShipGroup) -> Optional[Heading]:
         wind = self.game.conditions.weather.wind.at_0m
@@ -444,32 +433,6 @@ class GenericCarrierGenerator(GroundObjectGenerator):
 
 class CarrierGenerator(GenericCarrierGenerator):
     """Generator for CV(N) groups."""
-
-    def get_carrier_type(self, group: TheaterGroup) -> Type[ShipType]:
-        unit_type = super().get_carrier_type(group)
-        if self.game.settings.supercarrier:
-            unit_type = self.upgrade_to_supercarrier(unit_type, self.control_point.name)
-        return unit_type
-
-    @staticmethod
-    def upgrade_to_supercarrier(unit: Type[ShipType], name: str) -> Type[ShipType]:
-        if unit == Stennis:
-            if name == "CVN-71 Theodore Roosevelt":
-                return CVN_71
-            elif name == "CVN-72 Abraham Lincoln":
-                return CVN_72
-            elif name == "CVN-73 George Washington":
-                return CVN_73
-            elif name == "CVN-75 Harry S. Truman":
-                return CVN_75
-            elif name == "Carrier Strike Group 8":
-                return CVN_75
-            else:
-                return CVN_71
-        elif unit == KUZNECOW:
-            return CV_1143_5
-        else:
-            return unit
 
     def tacan_callsign(self) -> str:
         # TODO: Assign these properly.
